# src/config/swagger/swagger.yaml
swagger: "2.0"
info:
  title: Demo Publishing API
  version: 0.0.1
  description: The purpose of this API is to demonstrate the steps to integrate an existing Express project with OpenAPI
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Daniel Li
    url: http://danyll.com/
    email: dan@danyll.com
host: localhost:8080
basePath: /v1
consumes:
  - 'application/json'
produces:
  - 'text/plain; charset=utf-8'
  - 'application/json'
definitions:
...
  GenericError:
    type: object
    description: Generic Error Response
    properties:
      code:
        type: string
      message:
        type: string
    required:
    - code
    - message
...
responses:
  Success:
    description: Success Response
  InternalServerError:
    description: Internal Server Error Response
    schema:
      $ref: '#/definitions/GenericError'
  NotFound:
    description: Resource cannot be found
    schema:
      $ref: '#/definitions/GenericError'
...
paths:
...
    /posts/{postId}:
      x-swagger-router-controller: posts
      parameters:
      - name: postId
        in: path
        description: Post ID
        type: string
        required: true
      get:
        operationId: retrieve
        description: Retrieves a post by ID
        responses:
          '200':
            $ref: '#/responses/Success'
          '404':
            $ref: '#/responses/NotFound'
          '500':
            $ref: '#/responses/InternalServerError'
      delete:
        operationId: del
        description: Deletes a post by ID
        responses:
          '200':
            $ref: '#/responses/Success'
          '404':
            $ref: '#/responses/NotFound'
          '500':
            $ref: '#/responses/InternalServerError'
...